changes made to most files

+++ 0x02-functions_nested_loops/101-natural.c
@@ -13,7 +13,7 @@ return (0);
 void print_multiple(void)
 {
 int number;
-long unsigned int added;
+int added;
 for (number = 1023; number > 0; number--)
 {
 if ((number % 3) == 0)
@@ -21,5 +21,5 @@ added += number;
 else if ((number % 5) == 0)
 added += number;
 }
-printf("%lu\n", added);
+printf("%d\n", added);
 }
